fix: any type 제거 & 날짜 파싱 함수로 분리
