--- src/store/novelDetailData.ts
@@ -1,10 +1,0 @@
-export interface NovelDetailData {
-    id: string;
-    title: string;
-    authorList: string[];
-    category: string[];
-    coverImageUrl: string;
-    summary: string;
-    views: number;
-    likes: number;
-};
