+++ src/utils/parseDateTime.ts
@@ -1,0 +1,6 @@
+function formatDateTime(isoString: string): string {
+  const [date, time] = isoString.split('T');
+  return `${date} ${time}`;
+}
+
+export default formatDateTime;
